+++ tests/unit/test_librouteros.py
@@ -6,7 +6,7 @@ from unittest.mock import (
     Mock,
 )
 from librouteros import (
-    SYNC_DEFAULTS,
+    DEFAULTS,
     Api,
     connect,
     async_connect,
@@ -21,7 +21,7 @@ from librouteros.login import (
 
 def test_default_ssl_wrapper():
     """Assert that wrapper returns same object as it was called with."""
-    assert SYNC_DEFAULTS["ssl_wrapper"](int) is int
+    assert DEFAULTS["ssl_wrapper"](int) is int
 
 
 @pytest.mark.parametrize(
@@ -36,11 +36,11 @@ def test_default_ssl_wrapper():
     ),
 )
 def test_defaults(key, value):
-    assert SYNC_DEFAULTS[key] == value
+    assert DEFAULTS[key] == value
 
 
 def test_default_keys():
-    assert set(SYNC_DEFAULTS.keys()) == set(
+    assert set(DEFAULTS.keys()) == set(
         (
             "timeout",
             "port",